requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
altgraph==0.17.4
appdirs==1.4.4
atomicwrites==1.4.0
CacheControl==0.14.2
dateparser==1.2.1
Deprecated==1.2.18
fastapi==0.112.2
firebase-admin==6.6.0
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
future==1.0.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.2
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
httplib2==0.22.0
httptools==0.6.1
imageio-ffmpeg==0.6.0
jiter==0.8.2
jsonpointer==2.1
langchain==0.3.19
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.11
line-bot-sdk==3.16.0
mkl-service==2.4.0
moviepy==2.1.2
msgpack==1.1.0
oauthlib==3.2.2
openai==1.64.0
orjson==3.10.15
pdfminer.six==20231228
pdfplumber==0.11.4
pefile==2024.8.26
proglog==0.1.10
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1-modules==0.2.8
pycurl==7.45.2
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyls-spyder==0.4.0
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pypng==0.20220715.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pytz==2025.1
pywin32==305.1
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.5.1
starlette==0.38.2
tiktoken==0.9.0
typing_extensions==4.12.2
tzlocal==5.3
uritemplate==4.1.1
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
wheel==0.43.0
yarl==1.18.3
zstandard==0.23.0